WIP: allow conda to install whl files from a repo
A conda repo can include whl (Python wheel) files. These are extracted
into the package cache as noarch: python packages which can be linked
into environment.

Dependency information and other key metadata must be recorded in the
repodata which provides the wheel.

This is a rough version, various features of wheels are not supported
including entry-points and .data directories.