allow prefix records from wheel files
Do not reject prefix records that are created by transforming wheels in
the package cache.

These records will be for files that end with .whl.

In these files, hypens are allowed in the build string but not the
package name.

Follow up work will be needed to support packages whose name does include a hyphen.